STCOM-581 provide react-router as a peerDep
react-router and react-router-dom need to be provided by the platform as
dependencies for the platform's constituent modules to consume as
peerDependencies in order to guarantee that all modules are on the same
version. Modules may pass `<Link>` instances across module borders so we
need to be certain all modules agree on the same version of `<Link>` in
order to avoid crashing React.

Refs STCOM-581