strictly depend on react 16.13
`master` and some release branches had a caret dep on react, allowing
the installed version to float toward the latest minor release,
currently 16.13. This is unfortunate as it means our releases are
installing with a newer version of react than our test environments,
i.e. we were effectively releasing an untested build :scream:.

That'll be corrected in #554. This PR merely brings parity to
`#snapshot` and `#master`.