FOLIO-2738 temporarily disable courses
Tenant activation for mod-courses is failing, preventing all reference
builds using the platform-complete environment from building
successfully. This PR removes ui-courses, in order to prevent
mod-courses from being included in the build. It is anticipated that
this PR will soon be reverted as soon as the mod-courses activation
issue is resolved.

Refs FOLIO-2738