UICAL-161 dedupe style-loader with a resolutions entry
`ui-calendar` also has a hard-dep on `style-loader`, though we don't
know why, and trying to remove it is resulting in failing BigTest runs
in Jenkins, even though they succeed locally. Thing is, the duped
library is causing WYSIWYG panes to fail completely (e.g. the
ui-circulation circulation rules pane) so we need to get it out of
there. This is the way, at least for now.

Refs UICAL-161