FOLIO-3547 pin plugin deps to guarantee compatibility
Apps with loose (`*`) optional-deps on plugins allow incompatible
versions of those plugins to leak into the build. Even though the
platform's package.json may depend on `"ui-plugin-find-foo": "3.2.0"`,
if there is an app with a looser dependency that permits a later
version, then the later version will be pulled into the build.

If the later version is incompatible, e.g. because it has a peer
dependency on a stripes-acq-componets version that is not present in the
build, this can lead to corruption in the UI (e.g. missing
translations).

The following plugins are included by one or more apps with `*` deps and
therefore have `resolutions` entries here copied directly from their
strict `dependencies` version in order to guarantee that only compatible
versions are included:

* "@folio/plugin-find-contact": "3.1.1",
* "@folio/plugin-find-fund": "1.0.0",
* "@folio/plugin-find-instance": "6.1.6",
* "@folio/plugin-find-interface": "3.1.1",
* "@folio/plugin-find-organization": "3.1.1",
* "@folio/plugin-find-po-line": "3.1.1",
* "@folio/plugin-find-user": "6.1.0",

Refs FOLIO-3547