FOLIO-3602 provide additional shared third-party dependencies
Provide third-party dependencies that stripes and UI applications share
so we have a guarantee of compatibility. An identical version is a
requirement, enforced by React, in context Provider/Consumer
relationships, and while it's possible/probable that you'll get
compatible versions when both parties include deps like `^1.2.3`, it's
not technically a guarantee.

This also resolves several build warnings for apps that (rightly) expect
some of these things to be provided by the platform but in fact they
were only being provided directly by other apps and were being hoisted.

The following libraries are being added:

* final-form
* react-final-form
* final-form-arrays
* react-final-form-arrays
* redux-observable
* zustand

Refs FOLIO-3602