bump stripes-erm-components to v7.1.0
`stripes-erm-components` v7.1.0 has direct dependencies on `@folio/stripes-testing` and `@testing-library/react` in order to support shared test infrastructure. We don't really want that in the platform build, but treeshaking removes it from the bundle so it's not a problem in the end. Additional details in [ERM-2395](https://issues.folio.org/browse/ERM-2395) and https://github.com/folio-org/stripes-erm-components/pull/566.