UISAQCOMP-160 lock react-virtualized-auto-sizer to 1.0.7
Lock `react-virtualized-auto-sizer` to `v1.0.7` to preserve the CSS behavior
of `<Pane>` related to the "Actions" menu. RVAS >= `v1.0.8` substantially
reworked its internal structure, changing the way it interacts with
`<Pane>`, a bug described in detail in STCOM-1148.

Avoiding the update or changing our CSS are both viable approaches to
resolving the problem. Given `stripes-components` already takes the former
approach in its Orchid-compatible release (v11.0.4), we are taking the
same approach here.

Refs UISACQCOMP-160, STCOM-1148, STCOM-1146