FOLIO-3997 omit unnecessary favicons, sharp dependencies #2724
`favicons` and `sharp` were added to resolutions in #2644, probably to
avoid pulling `favicons` `7.1.5` in a build that was running without a
`yarn.lock` file before `@folio/stripes-webpack` `5.1.0` was published
(which similarly locks it to `v7.1.4`), but then dependabot updated in
PR #2645. Whoops. Bad dependabot. No biscuit! We depended on that
specific version because we wanted that specific version.

Refs [FOLIO-3997](https://folio-org.atlassian.net/browse/FOLIO-3997), [STRWEB-105](https://folio-org.atlassian.net/browse/STRWEB-105)