rsbuild/rspack POC
This was a fun diversion. It gets surprisingly close, but wacky CSS
prefixing kills it in the end. There's surely a way to resolve that, but
it's not immediately obvious what it is.

Grep for @@ to find meaningful changes/open questions in the config
files, which are largely copied verbatim from stripes-webpack.

Note @stripes/components/Icon will get hung up trying to configure its
`IconElement` so it's necessary to hack that function to just return sth
like `<span>{icon}</span>` until we figure out how to get SVG loading,
or whatever is breaking this, to unbreak.